+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## [0.0.2] - 15.04.2021
 
 * Add API calls for answering questions.
+
+## [0.1.0] - 19.04.2021
+
+* Add API calls for the API endpoint files.
+* Add further tests.